requirements.txt
appdirs==1.4.4
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
Django==3.0.8
django-filter==2.4.0
django-multiselectfield==0.1.12
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
mysql-connector-python==8.0.22
Pillow==8.1.0
psycopg2==2.8.6
pytz==2020.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
sqlparse==0.3.1
urllib3==1.26.3
virtualenv==20.0.28
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
